+++ test/reporter-cucumber-html/features/background/specs/attachments.feature.spec.ts
@@ -2,14 +2,22 @@ import { expect } from '@playwright/test';
 import { test } from '../../../check-report/fixtures';
 
 test('attachment in bg', async ({ feature }) => {
-  const background = feature.getBackground();
-  await expect(background.getSteps()).toContainText([
-    'Givenstep with attachment', // prettier-ignore
+  const scenario1 = feature.getScenario('scenario 1');
+  const scenario2 = feature.getScenario('scenario 2');
+
+  await expect(scenario1.getStepTitles()).toHaveText([
+    'Givenstep with attachment',
+    'GivenAction 1',
   ]);
-  await expect(background.getAttachments()).toHaveText([
+  await expect(scenario1.getAttachments()).toHaveText([
     'step with attachment|attachment for scenario scenario 1',
+  ]);
+
+  await expect(scenario2.getStepTitles()).toHaveText([
+    'Givenstep with attachment',
+    'GivenAction 2',
+  ]);
+  await expect(scenario2.getAttachments()).toHaveText([
     'step with attachment|attachment for scenario scenario 2',
   ]);
-  await expect(feature.getScenario('scenario 1').getAttachments()).toHaveCount(0);
-  await expect(feature.getScenario('scenario 2').getAttachments()).toHaveCount(0);
 });