+++ test/reporter-cucumber-html/features/before-all-hook/specs/named.feature.spec.ts
@@ -3,13 +3,10 @@ import { expect } from '@playwright/test';
 import { test } from '../../../check-report/fixtures';
 
 test('scenario 1', async ({ scenario }) => {
-  await expect(scenario.getSteps()).toContainText(
-    [
-      `Hook "my hook" failed: ${normalize('features/before-all-hook/steps.ts')}:`, // prettier-ignore
-      'GivenAction 1',
-      'Download trace',
-    ].filter(Boolean),
+  await expect(scenario.getStepTitles()).toContainText(
+    ['BeforeAll: my hook', 'GivenAction 1'].filter(Boolean),
   );
+  await expect(scenario.root).toContainText(normalize('features/before-all-hook/steps.ts'));
   await expect(scenario.getSteps('failed')).toHaveCount(1);
   await expect(scenario.getSteps('skipped')).toHaveCount(1);
   await expect(scenario.getErrors()).toContainText([`expect(true).toEqual(false)`]);