+++ test/reporter-cucumber-html/features/before-all-hook/specs/timeout.feature.spec.ts
@@ -2,10 +2,10 @@ import { expect } from '@playwright/test';
 import { test } from '../../../check-report/fixtures';
 
 test('scenario 1', async ({ scenario }) => {
-  await expect(scenario.getSteps()).toContainText([
+  await expect(scenario.getStepTitles()).toContainText([
     // Sometimes we still have 'BeforeAll Hooks' here, not 'my timeouted hook'.
     // It is when 'duration' != -1 -> no way to find exact timeouted item.
-    /Hook "(my timeouted hook|BeforeAll Hooks)" failed/, // prettier-ignore
+    /BeforeAll(?:: (my timeouted hook|BeforeAll Hooks))?/, // prettier-ignore
     'GivenAction 1',
   ]);
   await expect(scenario.getErrors()).toContainText([